requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
